+++ packages/common-all/src/nodev2.ts
@@ -772,10 +772,12 @@ export class NoteUtilsV2 {
   }
 
   static toLogObj(note: NotePropsV2) {
-    const { fname, id } = note;
+    const { fname, id, children, vault } = note;
     return {
       fname,
       id,
+      children,
+      vault,
     };
   }