+++ packages/dendron-11ty/libs/templates.js
@@ -1,0 +1,12 @@
+const { Liquid } = require('liquidjs');
+
+module.exports = (() => {
+  const liquidParser = new Liquid({
+    root: ['_includes'],
+    extname: '.liquid',
+    dynamicPartials: false,
+    strictFilters: true,
+  });
+
+  return liquidParser;
+})();