+++ packages/dendron-11ty/scripts/bump-minor.sh
@@ -1,0 +1,3 @@
+npm version minor
+npm publish
+git push