+++ packages/dendron-11ty/scripts/bump.sh
@@ -1,0 +1,3 @@
+npm version patch
+npm publish
+git push