+++ packages/dendron-11ty/scripts/link.sh
@@ -1,0 +1,7 @@
+# npm link @dendronhq/engine-server
+# npm link @dendronhq/common-server
+# npm link @dendronhq/common-all
+
+yarn link @dendronhq/engine-server
+yarn link @dendronhq/common-server
+yarn link @dendronhq/common-all