+++ packages/dendron-11ty/scripts/publish.sh
@@ -1,0 +1,4 @@
+./scripts/relink.sh
+git add .
+git commit -m "chore: update deps";
+./scripts/bump.sh