+++ packages/plugin-core/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.28.7-alpha.2](https://github.com/dendronhq/dendron/compare/v0.28.7-alpha.1...v0.28.7-alpha.2) (2021-02-14)
+
+**Note:** Version bump only for package @dendronhq/plugin-core
+
+## [0.28.7-alpha.1](https://github.com/dendronhq/dendron/compare/v0.28.7-alpha.0...v0.28.7-alpha.1) (2021-02-14)
+
+**Note:** Version bump only for package @dendronhq/plugin-core
+
 ## [0.28.7-alpha.0](https://github.com/dendronhq/dendron/compare/v0.28.6...v0.28.7-alpha.0) (2021-02-14)
 
 **Note:** Version bump only for package @dendronhq/plugin-core